[eclass] perl-module.eclass: default tests and parallel testing to ON.
This is a respectable change that needs much testing, and is a staging
test for ::gentoo to see how it works out.

Gentoo has only planned to phase in this kind of behaviour in, or after,
       EAPI6, so I'm chasing the concepts tail to make sure it will
       work.

Thus, the behaviour is now "On" by default in **all** EAPIs.

Things that break, please file bugs to have RESTRICT=test,
       PERL_RESTRICT=parallel-test or PERL_RESTRICT=network-test where
       applicable.

Also, there's a new user ENV control flag for those who want it,
  USER_PERL_RESTRICT.

  Adding "parallel-test" to this will disable parallel testing globally
  if its too much of a problem for you.

  Adding "network-test" will make Gentoo do their best to skip testing
  where network interaction is involved.

Enjoy.